refactor: rename generics module to generics_traits and update imports
