+++ crates/bevy_auto_plugin_shared/src/__private/util/generics_traits.rs
